requirements.txt
feedparser
jinja2
PyYAML
